[ Layers ] Mode Weight and Tensor name property to common
This pr move the Weight and Tensor name property to common property,
so that we can use other laysers too.

**Self evaluation:**
1. Build test:	 [X]Passed [ ]Failed [ ]Skipped
2. Run test:	 [X]Passed [ ]Failed [ ]Skipped
